+++ truffle/src/com.oracle.truffle.api.bytecode.test/src/com/oracle/truffle/api/bytecode/test/StackTraceTest.java
@@ -45,10 +45,13 @@ import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertSame;
+import static org.junit.Assert.assertThrows;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
+import static org.junit.Assume.assumeTrue;
 
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 
 import org.graalvm.polyglot.Context;
@@ -63,22 +66,29 @@ import org.junit.runners.Parameterized.Parameters;
 
 import com.oracle.truffle.api.CallTarget;
 import com.oracle.truffle.api.CompilerDirectives;
+import com.oracle.truffle.api.Truffle;
 import com.oracle.truffle.api.TruffleStackTrace;
 import com.oracle.truffle.api.TruffleStackTraceElement;
 import com.oracle.truffle.api.bytecode.BytecodeConfig;
+import com.oracle.truffle.api.bytecode.BytecodeLocal;
 import com.oracle.truffle.api.bytecode.BytecodeNode;
 import com.oracle.truffle.api.bytecode.BytecodeParser;
 import com.oracle.truffle.api.bytecode.BytecodeRootNode;
 import com.oracle.truffle.api.bytecode.BytecodeRootNodes;
 import com.oracle.truffle.api.bytecode.ConstantOperand;
+import com.oracle.truffle.api.bytecode.ContinuationResult;
 import com.oracle.truffle.api.bytecode.GenerateBytecode;
 import com.oracle.truffle.api.bytecode.GenerateBytecodeTestVariants;
 import com.oracle.truffle.api.bytecode.GenerateBytecodeTestVariants.Variant;
+import com.oracle.truffle.api.bytecode.Instruction;
 import com.oracle.truffle.api.bytecode.Operation;
 import com.oracle.truffle.api.dsl.Bind;
 import com.oracle.truffle.api.dsl.Specialization;
 import com.oracle.truffle.api.exception.AbstractTruffleException;
 import com.oracle.truffle.api.frame.FrameDescriptor;
+import com.oracle.truffle.api.frame.FrameInstance;
+import com.oracle.truffle.api.frame.FrameInstanceVisitor;
+import com.oracle.truffle.api.frame.VirtualFrame;
 import com.oracle.truffle.api.interop.ArityException;
 import com.oracle.truffle.api.interop.InteropLibrary;
 import com.oracle.truffle.api.interop.TruffleObject;
@@ -89,6 +99,7 @@ import com.oracle.truffle.api.library.ExportLibrary;
 import com.oracle.truffle.api.library.ExportMessage;
 import com.oracle.truffle.api.nodes.EncapsulatingNodeReference;
 import com.oracle.truffle.api.nodes.Node;
+import com.oracle.truffle.api.nodes.RootNode;
 import com.oracle.truffle.api.source.Source;
 
 @RunWith(Parameterized.class)
@@ -166,10 +177,7 @@ public class StackTraceTest extends AbstractInstructionTest {
                 Assert.fail();
             } catch (TestException e) {
                 List<TruffleStackTraceElement> elements = TruffleStackTrace.getStackTrace(e);
-                assertEquals(nodes.length, elements.size());
-                for (int i = 0; i < nodes.length; i++) {
-                    assertStackElement(elements.get(i), nodes[i], false);
-                }
+                assertStackElements(nodes, elements, "c.ThrowError", "c.Call", false);
             }
         }
     }
@@ -192,10 +200,7 @@ public class StackTraceTest extends AbstractInstructionTest {
                 Assert.fail();
             } catch (TestException e) {
                 List<TruffleStackTraceElement> elements = TruffleStackTrace.getStackTrace(e);
-                assertEquals(nodes.length, elements.size());
-                for (int i = 0; i < nodes.length; i++) {
-                    assertStackElement(elements.get(i), nodes[i], false);
-                }
+                assertStackElements(nodes, elements, "c.ThrowErrorBehindInterop", "c.Call", false);
             }
         }
     }
@@ -216,10 +221,7 @@ public class StackTraceTest extends AbstractInstructionTest {
 
         for (int repeat = 0; repeat < REPEATS; repeat++) {
             List<TruffleStackTraceElement> elements = (List<TruffleStackTraceElement>) outer.getCallTarget().call();
-            assertEquals(nodes.length, elements.size());
-            for (int i = 0; i < nodes.length; i++) {
-                assertStackElement(elements.get(i), nodes[i], false);
-            }
+            assertStackElements(nodes, elements, "c.CaptureStack", "c.Call", false);
         }
     }
 
@@ -244,15 +246,146 @@ public class StackTraceTest extends AbstractInstructionTest {
 
         for (int repeat = 0; repeat < REPEATS; repeat++) {
             List<TruffleStackTraceElement> elements = (List<TruffleStackTraceElement>) outer.getCallTarget().call();
-            assertEquals(nodes.length, elements.size());
-            for (int i = 0; i < nodes.length; i++) {
-                assertStackElement(elements.get(i), nodes[i], true);
-            }
+            assertStackElements(nodes, elements, "c.CaptureStack", "c.Call", true);
         }
     }
 
-    private void assertStackElement(TruffleStackTraceElement element, StackTraceTestRootNode target, boolean checkSources) {
-        assertSame(target.getCallTarget(), element.getTarget());
+    @Test
+    @SuppressWarnings("unchecked")
+    public void testValidateFrameInstances() {
+        int depth = run.depth;
+        StackTraceTestRootNode[] nodesConstant = new StackTraceTestRootNode[run.depth];
+        StackTraceTestRootNode[] nodes = chainCalls(depth, b -> {
+            b.beginRoot();
+            b.emitDummy();
+            b.emitValidateFrameInstances(run, nodesConstant, "c.Call");
+            b.endRoot();
+        }, true, false);
+        StackTraceTestRootNode outer = nodes[nodes.length - 1];
+        System.arraycopy(nodes, 0, nodesConstant, 0, depth);
+
+        for (int repeat = 0; repeat < REPEATS; repeat++) {
+            outer.getCallTarget().call();
+        }
+    }
+
+    @Test
+    @SuppressWarnings("unchecked")
+    public void testCaptureResumed() {
+        int dept = run.depth;
+        StackTraceTestRootNode[] nodes = chainResumes(dept, b -> {
+            b.beginRoot();
+            b.beginYield();
+            b.emitLoadNull();
+            b.endYield();
+            b.beginReturn();
+            b.emitCaptureStack();
+            b.endReturn();
+            b.endRoot();
+        }, false);
+        StackTraceTestRootNode outer = nodes[nodes.length - 1];
+
+        for (int repeat = 0; repeat < REPEATS; repeat++) {
+            ContinuationResult outerCont = (ContinuationResult) outer.getCallTarget().call();
+            List<TruffleStackTraceElement> elements = (List<TruffleStackTraceElement>) outerCont.continueWith(null);
+            assertStackElements(nodes, elements, "c.CaptureStack", "c.Resume", false);
+        }
+    }
+
+    @Test
+    @SuppressWarnings("unchecked")
+    public void testCaptureResumedWithSources() {
+        int dept = run.depth;
+        Source s = Source.newBuilder(BytecodeDSLTestLanguage.ID, "root0", "root0.txt").build();
+        StackTraceTestRootNode[] nodes = chainResumes(dept, b -> {
+            b.beginSource(s);
+            b.beginSourceSection(0, "root0".length());
+            b.beginRoot();
+            b.beginYield();
+            b.emitLoadNull();
+            b.endYield();
+            b.beginReturn();
+            b.emitCaptureStack();
+            b.endReturn();
+            b.endRoot();
+            b.endSourceSection();
+            b.endSource();
+        }, true);
+        StackTraceTestRootNode outer = nodes[nodes.length - 1];
+
+        for (int repeat = 0; repeat < REPEATS; repeat++) {
+            ContinuationResult outerCont = (ContinuationResult) outer.getCallTarget().call();
+            List<TruffleStackTraceElement> elements = (List<TruffleStackTraceElement>) outerCont.continueWith(null);
+            assertStackElements(nodes, elements, "c.CaptureStack", "c.Resume", true);
+        }
+    }
+
+    @Test
+    @SuppressWarnings("unchecked")
+    public void testValidateResumedFrameInstances() {
+        int depth = run.depth;
+        StackTraceTestRootNode[] nodesConstant = new StackTraceTestRootNode[run.depth];
+        StackTraceTestRootNode[] nodes = chainResumes(depth, b -> {
+            b.beginRoot();
+            b.beginYield();
+            b.emitLoadNull();
+            b.endYield();
+            b.emitValidateFrameInstances(run, nodesConstant, "c.Resume");
+            b.endRoot();
+        }, false);
+        StackTraceTestRootNode outer = nodes[nodes.length - 1];
+        System.arraycopy(nodes, 0, nodesConstant, 0, depth);
+
+        for (int repeat = 0; repeat < REPEATS; repeat++) {
+            ContinuationResult outerCont = (ContinuationResult) outer.getCallTarget().call();
+            outerCont.continueWith(null);
+        }
+    }
+
+    @Test
+    @SuppressWarnings("unchecked")
+    public void testCaptureWithLimit() {
+        int depth = run.depth;
+        assumeTrue("Cannot test stack trace limits with less than two frames in the trace.", depth >= 2);
+        int stackTraceLimit = depth / 2;
+        StackTraceTestRootNode[] nodes = chainCalls(depth, b -> {
+            b.beginRoot();
+            b.emitDummy();
+            b.beginReturn();
+            b.emitCaptureStackWithLimit(stackTraceLimit);
+            b.endReturn();
+            b.endRoot();
+        }, true, false);
+        StackTraceTestRootNode outer = nodes[nodes.length - 1];
+
+        // First, invoke as usual. The stack trace should consist of the top stackTraceLimit frames.
+        StackTraceTestRootNode[] topFrames = Arrays.copyOfRange(nodes, 0, stackTraceLimit);
+        for (int repeat = 0; repeat < REPEATS; repeat++) {
+            List<TruffleStackTraceElement> elements = (List<TruffleStackTraceElement>) outer.getCallTarget().call();
+            assertStackElements(topFrames, elements, "c.CaptureStackWithLimit", "c.Call", false);
+        }
+
+        // Next, mark the top (depth - stackTraceLimit) roots as internal so they do not "count".
+        for (int i = 0; i < depth - stackTraceLimit; i++) {
+            nodes[i].internal = true;
+        }
+        // Since they do not "count", we should capture every frame.
+        for (int repeat = 0; repeat < REPEATS; repeat++) {
+            List<TruffleStackTraceElement> elements = (List<TruffleStackTraceElement>) outer.getCallTarget().call();
+            assertStackElements(nodes, elements, "c.CaptureStackWithLimit", "c.Call", false);
+        }
+    }
+
+    private void assertStackElements(StackTraceTestRootNode[] nodes, List<TruffleStackTraceElement> elements, String firstElementInstruction, String chainedElementInstruction, boolean checkSources) {
+        assertEquals(nodes.length, elements.size());
+        assertStackElement(elements.get(0), nodes[0], firstElementInstruction, checkSources);
+        for (int i = 1; i < nodes.length; i++) {
+            assertStackElement(elements.get(i), nodes[i], chainedElementInstruction, checkSources);
+        }
+    }
+
+    private void assertStackElement(TruffleStackTraceElement element, StackTraceTestRootNode target, String instruction, boolean checkSources) {
+        assertEquals(target, run.interpreter.variant.cast(element.getTarget()));
         assertNotNull(element.getLocation());
         BytecodeNode bytecode = target.getBytecodeNode();
         if (run.interpreter.cached) {
@@ -261,7 +394,7 @@ public class StackTraceTest extends AbstractInstructionTest {
         } else {
             assertSame(bytecode, element.getLocation());
         }
-        assertEquals(bytecode.getInstructionsAsList().get(1).getBytecodeIndex(), element.getBytecodeIndex());
+        assertEquals(getBytecodeIndexOfFirstInstruction(bytecode, instruction), element.getBytecodeIndex());
 
         Object interopObject = element.getGuestObject();
         InteropLibrary lib = InteropLibrary.getFactory().create(interopObject);
@@ -276,7 +409,16 @@ public class StackTraceTest extends AbstractInstructionTest {
         } catch (UnsupportedMessageException ex) {
             fail("Interop object could not receive message: " + ex);
         }
+    }
 
+    private static int getBytecodeIndexOfFirstInstruction(BytecodeNode bytecode, String instructionName) {
+        for (Instruction i : bytecode.getInstructions()) {
+            if (instructionName.equals(i.getName())) {
+                return i.getBytecodeIndex();
+            }
+        }
+        fail("No instruction found matching instruction name " + instructionName);
+        return -1;
     }
 
     @Test
@@ -309,6 +451,52 @@ public class StackTraceTest extends AbstractInstructionTest {
         assertNull(element.getLocation());
         assertEquals(-1, element.getBytecodeIndex());
         assertNull(BytecodeNode.get(element));
+    }
+
+    @Test
+    public void testBadLocationFrameInstance() {
+        try (Context c = BytecodeDSLTestLanguage.createPolyglotContextWithCompilationDisabled()) {
+            CallTarget collectBytecodeNodes = new RootNode(null) {
+                @Override
+                public Object execute(VirtualFrame frame) {
+                    List<BytecodeNode> bytecodeNodes = new ArrayList<>();
+                    Truffle.getRuntime().iterateFrames(f -> {
+                        bytecodeNodes.add(BytecodeNode.get(f));
+                        return null;
+                    });
+                    return bytecodeNodes;
+                }
+            }.getCallTarget();
+
+            StackTraceTestRootNode caller = parse(b -> {
+                b.beginRoot();
+                b.emitCallWithBadLocation(collectBytecodeNodes);
+                b.endRoot();
+            });
+
+            assertThrows(AssertionError.class, () -> caller.getCallTarget().call());
+        }
+    }
+
+    @Test
+    public void testBadLocationTruffleStackTraceElement() {
+        try (Context c = BytecodeDSLTestLanguage.createPolyglotContextWithCompilationDisabled()) {
+            StackTraceTestRootNode capturesStack = parse(b -> {
+                b.beginRoot();
+                b.beginReturn();
+                b.emitCaptureStack();
+                b.endReturn();
+                b.endRoot();
+            });
+
+            StackTraceTestRootNode caller = parse(b -> {
+                b.beginRoot();
+                b.emitCallWithBadLocation(capturesStack.getCallTarget());
+                b.endRoot();
+            });
+
+            assertThrows(AssertionError.class, () -> caller.getCallTarget().call());
+        }
     }
 
     private StackTraceTestRootNode[] chainCalls(int depth, BytecodeParser<StackTraceTestRootNodeBuilder> innerParser, boolean includeLocation, boolean includeSources) {
@@ -345,15 +533,64 @@ public class StackTraceTest extends AbstractInstructionTest {
         return nodes;
     }
 
+    private StackTraceTestRootNode[] chainResumes(int depth, BytecodeParser<StackTraceTestRootNodeBuilder> innerParser, boolean includeSources) {
+        // Constructs a chain of methods that call the previous method, then yield, and then resume
+        // the previous method.
+        StackTraceTestRootNode[] nodes = new StackTraceTestRootNode[depth];
+        nodes[0] = parse(innerParser);
+        nodes[0].setName("root0");
+        for (int i = 1; i < depth; i++) {
+            int index = i;
+            String name = "root" + i;
+            Source s = includeSources ? Source.newBuilder(BytecodeDSLTestLanguage.ID, name, name + ".txt").build() : null;
+            nodes[i] = parse(b -> {
+                // x = prev_root()
+                // yield
+                // resume(x)
+                if (includeSources) {
+                    b.beginSource(s);
+                    b.beginSourceSection(0, name.length());
+                }
+                b.beginRoot();
+                BytecodeLocal calleeContinuation = b.createLocal();
+                b.beginStoreLocal(calleeContinuation);
+                b.emitCall(nodes[index - 1].getCallTarget());
+                b.endStoreLocal();
+
+                b.beginYield();
+                b.emitLoadConstant(index);
+                b.endYield();
+
+                b.beginReturn();
+                b.beginResume();
+                b.emitLoadLocal(calleeContinuation);
+                b.emitLoadNull();
+                b.endResume();
+                b.endReturn();
+                b.endRoot().depth = index;
+                if (includeSources) {
+                    b.endSourceSection();
+                    b.endSource();
+                }
+            });
+            nodes[i].setName(name);
+        }
+        return nodes;
+    }
+
     @GenerateBytecodeTestVariants({
                     @Variant(suffix = "CachedDefault", configuration = //
-                    @GenerateBytecode(languageClass = BytecodeDSLTestLanguage.class, storeBytecodeIndexInFrame = false, enableUncachedInterpreter = false, boxingEliminationTypes = {int.class})),
+                    @GenerateBytecode(languageClass = BytecodeDSLTestLanguage.class, storeBytecodeIndexInFrame = false, enableUncachedInterpreter = false, enableYield = true, boxingEliminationTypes = {
+                                    int.class})),
                     @Variant(suffix = "UncachedDefault", configuration = //
-                    @GenerateBytecode(languageClass = BytecodeDSLTestLanguage.class, storeBytecodeIndexInFrame = false, enableUncachedInterpreter = true, boxingEliminationTypes = {int.class})),
+                    @GenerateBytecode(languageClass = BytecodeDSLTestLanguage.class, storeBytecodeIndexInFrame = false, enableUncachedInterpreter = true, enableYield = true, boxingEliminationTypes = {
+                                    int.class})),
                     @Variant(suffix = "CachedBciInFrame", configuration = //
-                    @GenerateBytecode(languageClass = BytecodeDSLTestLanguage.class, storeBytecodeIndexInFrame = true, enableUncachedInterpreter = false, boxingEliminationTypes = {int.class})),
+                    @GenerateBytecode(languageClass = BytecodeDSLTestLanguage.class, storeBytecodeIndexInFrame = true, enableUncachedInterpreter = false, enableYield = true, boxingEliminationTypes = {
+                                    int.class})),
                     @Variant(suffix = "UncachedBciInFrame", configuration = //
-                    @GenerateBytecode(languageClass = BytecodeDSLTestLanguage.class, storeBytecodeIndexInFrame = true, enableUncachedInterpreter = true, boxingEliminationTypes = {int.class}))
+                    @GenerateBytecode(languageClass = BytecodeDSLTestLanguage.class, storeBytecodeIndexInFrame = true, enableUncachedInterpreter = true, enableYield = true, boxingEliminationTypes = {
+                                    int.class}))
     })
     public abstract static class StackTraceTestRootNode extends DebugBytecodeRootNode implements BytecodeRootNode {
 
@@ -363,6 +600,7 @@ public class StackTraceTest extends AbstractInstructionTest {
 
         public String name;
         public int depth;
+        public boolean internal = false;
 
         public void setName(String name) {
             this.name = name;
@@ -371,6 +609,11 @@ public class StackTraceTest extends AbstractInstructionTest {
         @Override
         public String getName() {
             return name;
+        }
+
+        @Override
+        public boolean isInternal() {
+            return internal;
         }
 
         @Override
@@ -402,6 +645,25 @@ public class StackTraceTest extends AbstractInstructionTest {
             @Specialization
             static Object doDefault(CallTarget target) {
                 return target.call((Node) null);
+            }
+        }
+
+        @Operation(storeBytecodeIndex = true)
+        @ConstantOperand(type = CallTarget.class)
+        static final class CallWithBadLocation {
+            @Specialization
+            static Object doDefault(CallTarget target, @Bind StackTraceTestRootNode root) {
+                // This is incorrect. The tests using this operation exert paths that should trigger
+                // assertions.
+                return target.call(root);
+            }
+        }
+
+        @Operation(storeBytecodeIndex = true)
+        static final class Resume {
+            @Specialization
+            static Object doDefault(ContinuationResult cont, Object resumeValue, @Bind Node node) {
+                return cont.getContinuationCallTarget().call(node, cont.getFrame(), resumeValue);
             }
         }
 
@@ -445,6 +707,75 @@ public class StackTraceTest extends AbstractInstructionTest {
                 return TruffleStackTrace.getStackTrace(ex);
             }
         }
+
+        @Operation(storeBytecodeIndex = true)
+        @ConstantOperand(type = int.class, name = "stackTraceLimit")
+        static final class CaptureStackWithLimit {
+
+            @Specialization
+            static Object doDefault(int stackTraceLimit, @Bind Node node) {
+                TestException ex = new TestException(node, stackTraceLimit);
+                return TruffleStackTrace.getStackTrace(ex);
+            }
+        }
+
+        @SuppressWarnings("truffle-interpreted-performance")
+        @Operation(storeBytecodeIndex = true)
+        // The parser should populate this array after parsing. This is a hack to expose the roots.
+        @ConstantOperand(type = Run.class, name = "run")
+        @ConstantOperand(type = StackTraceTestRootNode[].class, name = "rootNodes")
+        @ConstantOperand(type = String.class, name = "chainedInstruction")
+        static final class ValidateFrameInstances {
+
+            @Specialization
+            static void doDefault(Run run, StackTraceTestRootNode[] rootNodes, String chainedInstruction) {
+                FrameInstanceValidator visitor = new FrameInstanceValidator(run, rootNodes, chainedInstruction);
+                Truffle.getRuntime().iterateFrames(visitor);
+                visitor.checkFinished();
+            }
+        }
+
+        static final class FrameInstanceValidator implements FrameInstanceVisitor<Void> {
+            public final Run run;
+            public final StackTraceTestRootNode[] targets;
+            public final String chainedFrameInstruction;
+            public int index = 0;
+
+            FrameInstanceValidator(Run run, StackTraceTestRootNode[] targets, String chainedFrameInstruction) {
+                this.run = run;
+                this.targets = targets;
+                this.chainedFrameInstruction = chainedFrameInstruction;
+            }
+
+            public Void visitFrame(FrameInstance frameInstance) {
+                if (index >= targets.length) {
+                    fail("Visited too many frames.");
+                }
+                StackTraceTestRootNode target = targets[index];
+
+                assertEquals(target, run.interpreter.variant.cast(frameInstance.getCallTarget()));
+                BytecodeNode bytecode = target.getBytecodeNode();
+                if (index != 0) {
+                    assertNotNull(frameInstance.getCallNode());
+                    if (run.interpreter.cached) {
+                        assertSame(bytecode, BytecodeNode.get(frameInstance.getCallNode()));
+                        assertSame(bytecode, BytecodeNode.get(frameInstance));
+                    } else {
+                        assertSame(bytecode, frameInstance.getCallNode());
+                    }
+                    assertEquals(getBytecodeIndexOfFirstInstruction(bytecode, chainedFrameInstruction), frameInstance.getBytecodeIndex());
+                }
+                index++;
+                return null;
+            }
+
+            public void checkFinished() {
+                if (index != targets.length) {
+                    fail("Did not visit every frame. Index is " + index + " and there are " + targets.length + " expected frames.");
+                }
+            }
+
+        }
     }
 
     @ExportLibrary(InteropLibrary.class)
@@ -469,6 +800,10 @@ public class StackTraceTest extends AbstractInstructionTest {
 
         TestException(Node location) {
             super(resolveLocation(location));
+        }
+
+        TestException(Node location, int stackTraceElementLimit) {
+            super(null, null, stackTraceElementLimit, resolveLocation(location));
         }
 
         private static Node resolveLocation(Node location) {